Revert "Исправлены юнит-тесты для методов delete(Model model)"
