+++ UnitTestOfTimetableOfClasses/UT_Insert_CInstitute.cs
@@ -2,24 +2,24 @@
 //using Microsoft.VisualStudio.TestTools.UnitTesting;
 //using LibOfTimetableOfClasses;
 
-namespace UnitTestOfTimetableOfClasses
-{
-	[TestClass]
-	public class UT_Insert_CInstitute
-	{
-		[TestMethod]
-		public void Task_515_1() //Добавление в пустую таблицу
-		{
-            //arrange 
-            Controllers.CInstitute.Select().Clear();
-            MInstitute inst = new MInstitute("ИАСТ", "Институт Автоматизированных систем и технологий", "Лустгартен Ю.Л.");
-			bool expected = true;
-            //act
-            bool actual = Controllers.CInstitute.Insert(inst);
-			//assert
-			Assert.AreEqual(expected, actual);
-		}
 
+//namespace UnitTestOfTimetableOfClasses
+//{
+//	[TestClass]
+//	public class UT_Insert_CInstitute
+//	{
+//		[TestMethod]
+//		public void Task_515_1() //Добавление в пустую таблицу
+//		{
+//			//arrange 
+//			MInstitute inst = new MInstitute("ИАСТ", "Институт Автоматизированных систем и технологий", "Лустгартен Ю.Л.");
+//			bool expected = true;
+//			//act
+//			CInstitute I = new CInstitute();
+//			bool actual = I.Insert(inst);
+//			//assert
+//			Assert.AreEqual(expected, actual);
+//		}
 
 //		[TestMethod]
 //		public void Task_515_2() //Повторяющийся атрибут "Краткое название"
@@ -64,5 +64,5 @@ namespace UnitTestOfTimetableOfClasses
 //			//assert
 //			Assert.AreEqual(expected, actual);
 //		}
-   	}
-  }
+//	}
+//}
