+++ UnitTestOfTimetableOfClasses/UnitTestOfTimetableOfClasses.csproj
@@ -43,7 +43,7 @@
   <ItemGroup>
     <Compile Include="UT_Delete_CEnclosures.cs" />
     <Compile Include="UT_Delete_CUniversity.cs" />
-    <Compile Include="UT_Delete_CTitle.cs" />
+    <Compile Include="UI_Delete_CTitle.cs" />
     <Compile Include="UT_Delete_CInstitute.cs" />
     <Compile Include="UT_Insert_CInstitute.cs" />
     <Compile Include="UT_Update_CInstitute.cs" />
@@ -94,9 +94,6 @@
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Data" />
-    <Reference Include="System.ValueTuple, Version=4.0.3.0, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51, processorArchitecture=MSIL">
-      <HintPath>..\packages\System.ValueTuple.4.5.0\lib\netstandard1.0\System.ValueTuple.dll</HintPath>
-    </Reference>
   </ItemGroup>
   <Import Project="$(VSToolsPath)\TeamTest\Microsoft.TestTools.targets" Condition="Exists('$(VSToolsPath)\TeamTest\Microsoft.TestTools.targets')" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
