+++ UnitTestOfTimetableOfClasses/packages.config
@@ -2,5 +2,4 @@
 <packages>
   <package id="MSTest.TestAdapter" version="1.4.0" targetFramework="net45" />
   <package id="MSTest.TestFramework" version="1.4.0" targetFramework="net45" />
-  <package id="System.ValueTuple" version="4.5.0" targetFramework="net45" />
 </packages>
